move constant and delete redundant file
